Codes refactoring.
