+++ faostat-api-core/src/main/java/org/fao/faostat/api/core/FAOSTATAPICore.java
@@ -342,15 +342,13 @@
 package org.fao.faostat.api.core;
 
 import org.fao.faostat.api.core.beans.DatasourceBean;
+import org.fao.faostat.api.core.beans.FAOSTATIterable;
 import org.fao.faostat.api.core.beans.MetadataBean;
 import org.fao.faostat.api.core.beans.OutputBean;
 import org.fao.faostat.api.core.constants.QUERIES;
 import org.fao.faostat.api.core.jdbc.JDBCIterable;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 /**
  * @author <a href="mailto:guido.barbaglia@gmail.com">Guido Barbaglia</a>
@@ -507,6 +505,278 @@ public class FAOSTATAPICore {
             throw new Exception(log.toString());
         }
 
+    }
+
+    public OutputBean queryCodes(DatasourceBean datasourceBean, MetadataBean metadataBean) throws Exception {
+
+        /* Logs. */
+        StringBuilder log = new StringBuilder();
+
+
+        try {
+
+            /* Statistics. */
+            log.append("FAOSTATAPICore\t").append("initiate statistics...").append("\n");
+            long t0 = System.currentTimeMillis();
+
+            /* Get domain's dimensions. */
+            List<List<Map<String, Object>>> dimensions = getDomainDimensions("dimensions", datasourceBean, metadataBean);
+
+            /* Organize data in an object. */
+            final List<Map<String, Object>> structuredDimensions = organizeDomainDimensions(dimensions);
+
+            /* Get the dimension of interest. */
+            Map<String, Object> dimension = null;
+            String dimensionCode = metadataBean.getProcedureParameters().get("id").toString();
+            for (Map<String, Object> m : structuredDimensions) {
+                if (m.get("id").toString().equalsIgnoreCase(dimensionCode)) {
+                    dimension = m;
+                    break;
+                }
+                ArrayList<Map<String, Object>> subdimensions = (ArrayList<Map<String, Object>>) m.get("subdimensions");
+                for (Map<String, Object> subm : subdimensions) {
+                    if (subm.get("id").toString().equalsIgnoreCase(dimensionCode)) {
+                        dimension = subm;
+                        break;
+                    }
+                }
+            }
+
+            /* Prepare output. */
+            List<Map<String, Object>> codes = new ArrayList<>();
+            Map<String, Object> row;
+
+            /* Get subdimensions. */
+            ArrayList<Map<String, Object>> subDimensions = (ArrayList<Map<String, Object>>) dimension.get("subdimensions");
+
+            if (subDimensions != null) {
+
+                /* Fetch codes for each sub-dimension. */
+                for (Map<String, Object> m : subDimensions) {
+
+                    /* Define options to fetch codes. */
+                    MetadataBean subDimensionOptions = new MetadataBean();
+                    subDimensionOptions.addParameter("domain_code", metadataBean.getProcedureParameters().get("domain_code"));
+                    subDimensionOptions.addParameter("lang", metadataBean.getProcedureParameters().get("lang"));
+                    subDimensionOptions.addParameter("dimension", m.get("ListBoxNo").toString());
+                    subDimensionOptions.addParameter("subdimension", m.get("TabOrder").toString());
+                    JDBCIterable subDimensionIterable = getJDBCIterable("codes", datasourceBean, subDimensionOptions);
+
+                    /* Iterate over codes. */
+                    while (subDimensionIterable.hasNext()) {
+                        row = createCode(subDimensionIterable.nextMap());
+                        row.put("subdimension_id", m.get("TabName").toString().replace(" ", "").toLowerCase());
+                        row.put("subdimension_ord", m.get("TabOrder").toString());
+                        row.put("subdimension_label", m.get("TabName").toString());
+                        log.append("\t\tcontains ").append(row.get("code").toString().toUpperCase()).append("? ").append(Arrays.asList(metadataBean.getBlackList()).contains(row.get("code").toString().toUpperCase())).append("\n");
+                        if (isAdmissibleCode(row, metadataBean))
+                            codes.add(row);
+                    }
+
+                }
+
+            } else {
+
+                /* Fetch codes. */
+                MetadataBean subDimensionOptions = new MetadataBean();
+                subDimensionOptions.addParameter("domain_code", metadataBean.getProcedureParameters().get("domain_code"));
+                subDimensionOptions.addParameter("lang", metadataBean.getProcedureParameters().get("lang"));
+                subDimensionOptions.addParameter("dimension", dimension.get("ListBoxNo").toString());
+                subDimensionOptions.addParameter("subdimension", dimension.get("TabOrder").toString());
+                JDBCIterable subDimensionIterable = getJDBCIterable("codes", datasourceBean, subDimensionOptions);
+                while (subDimensionIterable.hasNext()) {
+                    row = createCode(subDimensionIterable.nextMap());
+                    if (isAdmissibleCode(row, metadataBean))
+                        codes.add(row);
+                }
+
+            }
+
+            /* Add children. */
+            for (Map<String, Object> c1 : codes) {
+                if (c1.get("parent") != null) {
+                    for (Map<String, Object> c2 : codes) {
+                        if (c2.get("code").toString().equalsIgnoreCase(c1.get("parent").toString())) {
+                            ((ArrayList<Map<String, Object>>)c2.get("children")).add(c1);
+                        }
+                    }
+                }
+            }
+
+            /* Group subdimensions, if required. */
+            final List<Map<String, Object>> output = new ArrayList<>();
+            if (Boolean.parseBoolean(metadataBean.getProcedureParameters().get("group_subdimensions").toString())) {
+
+                /* Prepare output. */
+                for (Map<String, Object> m : subDimensions) {
+                    Map<String, Object> subdimension = new HashMap<>();
+                    subdimension.put("metadata", new HashMap<String, Object>());
+                    subdimension.put("data", new ArrayList<HashMap<String, Object>>());
+                    output.add(subdimension);
+                }
+
+                /* Group codes. */
+                String current_id = codes.get(0).get("subdimension_id").toString();
+                String current_ord = codes.get(0).get("subdimension_ord").toString();
+                String current_label = codes.get(0).get("subdimension_label").toString();
+                int current_idx = 0;
+                for (int i = 0; i < codes.size(); i += 1) {
+
+                    /* Current code. */
+                    Map<String, Object> c = codes.get(i);
+
+                    if (c.get("subdimension_id").toString().equalsIgnoreCase(current_id)) {
+                        ((ArrayList<Map<String, Object>>)output.get(current_idx).get("data")).add(c);
+                    } else {
+
+                        /* Write metadata. */
+                        ((HashMap<String, Object>)output.get(current_idx).get("metadata")).put("id", current_id);
+                        ((HashMap<String, Object>)output.get(current_idx).get("metadata")).put("ord", current_ord);
+                        ((HashMap<String, Object>)output.get(current_idx).get("metadata")).put("label", current_label);
+
+                        /* Reset parameters. */
+                        current_id = codes.get(i).get("subdimension_id").toString();
+                        current_ord = codes.get(i).get("subdimension_ord").toString();
+                        current_label = codes.get(i).get("subdimension_label").toString();
+                        current_idx += 1;
+                        ((ArrayList<Map<String, Object>>)output.get(current_idx).get("data")).add(c);
+
+                    }
+                }
+
+                /* Write metadata for the last iteration. */
+                ((HashMap<String, Object>)output.get(current_idx).get("metadata")).put("id", current_id);
+                ((HashMap<String, Object>)output.get(current_idx).get("metadata")).put("ord", current_ord);
+                ((HashMap<String, Object>)output.get(current_idx).get("metadata")).put("label", current_label);
+
+            } else {
+                output.addAll(codes);
+            }
+
+//            throw new Exception(output.toString());
+
+            /* Initiate output. */
+            log.append("FAOSTATAPICore\t").append("initiate out...").append("\n");
+            OutputBean out = new OutputBean(new FAOSTATIterable(output));
+            log.append("FAOSTATAPICore\t").append("initiate out: done").append("\n");
+
+            /* Add metadata. */
+            log.append("FAOSTATAPICore\t").append("add metadata...").append("\n");
+            out.setMetadata(metadataBean);
+
+            /* Statistics. */
+            long tf = System.currentTimeMillis();
+            log.append("FAOSTATAPICore\t").append("set statistics...").append("\n");
+            out.getMetadata().setProcessingTime(tf - t0);
+
+            /* Return output. */
+            log.append("FAOSTATAPICore\t").append("return output...").append("\n");
+            return out;
+
+        } catch (Exception e) {
+            throw new Exception(log.toString());
+        }
+
+    }
+
+    private Map<String, Object> createCode(Map<String, Object> dbRow) {
+        Map<String, Object> code = new HashMap<>();
+        code.put("code", dbRow.get("Code"));
+        code.put("label", dbRow.get("Label"));
+        code.put("ord", dbRow.get("Order"));
+        code.put("parent", dbRow.get("Parent"));
+        code.put("description", "TODO");
+        code.put("aggregate_type", dbRow.get("AggregateType"));
+        code.put("children", new ArrayList<Map<String, Object>>());
+        return code;
+    }
+
+    private boolean isAdmissibleCode(Map<String, Object> row, MetadataBean o) {
+
+        /* Check wheter the code is a list. */
+        if (row.get("aggregate_type").equals(">")) {
+            return Boolean.parseBoolean(o.getProcedureParameters().get("show_lists").toString());
+        } else {
+
+            /* Check the blacklist. */
+            if (o.getBlackList() != null && o.getBlackList().size() > 0) {
+                if (o.getBlackList().contains(row.get("code").toString())) {
+                    return false;
+                }
+            }
+
+            /* Check the whitelist. */
+            if (o.getWhiteList() != null && o.getWhiteList().size() > 0) {
+                if (!o.getWhiteList().contains(row.get("code").toString())) {
+                    return false;
+                }
+            }
+
+        }
+
+        /* Return. */
+        return true;
+
+    }
+
+    private List<List<Map<String, Object>>> getDomainDimensions(String queryCode, DatasourceBean datasourceBean, MetadataBean o) throws Exception {
+
+        /* Query the DB. */
+        JDBCIterable i = getJDBCIterable(queryCode, datasourceBean, o);
+
+        /* Store the original result. */
+        List<Map<String, Object>> l = new ArrayList<>();
+        while (i.hasNext()) {
+            l.add(i.nextMap());
+        }
+
+        /* Initiate variables. */
+        List<List<Map<String, Object>>> dimensions = new ArrayList<>();
+
+        /* Create groups. */
+        List<Map<String, Object>> groups = new ArrayList<>();
+        String current = "1";
+        for (Map<String, Object> m : l) {
+            m.put("id", m.get("TabName").toString().replaceAll(" ", "").toLowerCase());
+            if (m.get("ListBoxNo").toString().equalsIgnoreCase(current)) {
+                groups.add(m);
+            } else {
+                dimensions.add(groups);
+                current = m.get("ListBoxNo").toString();
+                groups = new ArrayList<>();
+                groups.add(m);
+            }
+        }
+        dimensions.add(groups);
+
+        /* Return output. */
+        return dimensions;
+
+    }
+
+    private List<Map<String, Object>> organizeDomainDimensions(List<List<Map<String, Object>>> dimensions) {
+        List<Map<String, Object>> out = new ArrayList<>();
+        List<String> idsBuffer = new ArrayList<>();
+        for (List<Map<String, Object>> dimension : dimensions) {
+            Map<String, Object> tmp = new HashMap<>();
+            if (!idsBuffer.contains(dimension.get(0).get("VarTypeGroup") + "group")) {
+                idsBuffer.add(dimension.get(0).get("VarTypeGroup") + "group");
+                tmp.put("id", dimension.get(0).get("VarTypeGroup") + "group");
+            } else {
+                tmp.put("id", dimension.get(0).get("VarTypeGroup") + "group2");
+            }
+            tmp.put("ord", dimension.get(0).get("ListBoxNo"));
+            tmp.put("label", "TODO");
+            tmp.put("description", "TODO");
+            tmp.put("parameter", "@List" + dimension.get(0).get("ListBoxNo") + "Codes");
+            tmp.put("href", "/codes/" + dimension.get(0).get("VarTypeGroup") + "group/");
+            tmp.put("subdimensions", new ArrayList<Map<String, Object>>());
+            for (Map<String, Object> subdimension : dimension) {
+                ((ArrayList)tmp.get("subdimensions")).add(subdimension);
+            }
+            out.add(tmp);
+        }
+        return out;
     }
 
     private boolean isAdmissibleDBRow(Map<String, Object> row, MetadataBean o) {
